feat(apps): move git status from devtools to per-app tab
