+++ .changelog/v1.21.x.md
@@ -12,6 +12,10 @@ GSD tab one-click agent spawning and Claude Code launcher.
 - **Open Claude Code button** — new button in the GSD setup guide header launches Claude Code directly in the app's directory for interactive use
 - **`POST /api/apps/:id/open-claude` endpoint** — spawns Claude Code as a detached process in the app's repo path
 
+## Improvements
+
+- **Git status moved to per-app tab** — git status is now a tab within each app's detail view (`/apps/:appId/git`) instead of a standalone cross-app page under Dev Tools, making it contextual to the app you're working on
+
 ## Fixes
 
 - **PortOS self-restart no longer errors** — restarting PortOS from the dashboard now shows a "Restarting PortOS..." loading toast, polls for server recovery, and auto-reloads the page when ready (previously showed an error because the server died before responding)
