+++ client/src/App.jsx
@@ -26,7 +26,6 @@ const AIProviders = lazy(() => import('./pages/AIProviders'));
 const HistoryPage = lazy(() => import('./pages/DevTools').then(m => ({ default: m.HistoryPage })));
 const RunsHistoryPage = lazy(() => import('./pages/DevTools').then(m => ({ default: m.RunsHistoryPage })));
 const RunnerPage = lazy(() => import('./pages/DevTools').then(m => ({ default: m.RunnerPage })));
-const GitPage = lazy(() => import('./pages/DevTools').then(m => ({ default: m.GitPage })));
 const UsagePage = lazy(() => import('./pages/DevTools').then(m => ({ default: m.UsagePage })));
 const ProcessesPage = lazy(() => import('./pages/DevTools').then(m => ({ default: m.ProcessesPage })));
 const AgentsPage = lazy(() => import('./pages/DevTools').then(m => ({ default: m.AgentsPage })));
@@ -51,7 +50,6 @@ export default function App() {
           <Route path="devtools/history" element={<HistoryPage />} />
           <Route path="devtools/runs" element={<RunsHistoryPage />} />
           <Route path="devtools/runner" element={<RunnerPage />} />
-          <Route path="devtools/git" element={<GitPage />} />
           <Route path="devtools/usage" element={<UsagePage />} />
           <Route path="devtools/processes" element={<ProcessesPage />} />
           <Route path="devtools/agents" element={<AgentsPage />} />
