+++ client/src/components/Layout.jsx
@@ -13,7 +13,6 @@ import {
   History,
   Code2,
   Activity,
-  GitBranch,
   BarChart3,
   Cpu,
   Wrench,
@@ -107,7 +106,6 @@ const navItems = [
       { href: '//:5560', label: 'Autofixer', icon: Wrench, external: true, dynamicHost: true },
       { to: '/browser', label: 'Browser', icon: Globe },
       { to: '/devtools/runner', label: 'Code', icon: Code2 },
-      { to: '/devtools/git', label: 'Git Status', icon: GitBranch },
       { to: '/devtools/history', label: 'History', icon: History },
       { to: '/devtools/jira', label: 'JIRA', icon: Ticket },
       { to: '/devtools/processes', label: 'Processes', icon: Activity },
