+++ client/src/components/apps/constants.js
@@ -3,5 +3,6 @@ export const APP_DETAIL_TABS = [
   { id: 'tasks', label: 'Tasks' },
   { id: 'automation', label: 'Automation' },
   { id: 'documents', label: 'Documents' },
+  { id: 'git', label: 'Git' },
   { id: 'gsd', label: 'GSD' }
 ];
